Fix redeclared classname for some migrations and add conditions if creating product_user table

+++ modules/Ecommerce/Database/Migrations/2022_09_19_032626_create_product_user_table.php
@@ -13,19 +13,21 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('product_user', function (Blueprint $table) {
-            $table->id();
-            $table
-                ->foreignId('product_id')
-                ->constrained(config('lunar.database.table_prefix').'products')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-            $table
-                ->foreignId('user_id')
-                ->constrained()
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+        if (!Schema::hasTable('product_user')) {
+            Schema::create('product_user', function (Blueprint $table) {
+                $table->id();
+                $table
+                    ->foreignId('product_id')
+                    ->constrained(config('lunar.database.table_prefix').'products')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+                $table
+                    ->foreignId('user_id')
+                    ->constrained()
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
